feat: add range slider with custom scale
